requirements.txt
gym==0.17.1
gym_retro==0.7.0
keras
tensorflow
